requirements.txt
mem0ai
google-genai
streamlit
chainlit
ipykernel
Pinecone
langchain
unstructured[all]
pymupdf
pytesseract
pillow